c/inodecache: optimize read locks
When attempting to obtain a shared lock bail out if the .lock
file does not exist. Shared lock is used for reading, however
if lock file does not exist neither does the data file.
Therefore on a read code path it's OK to skip `O_TRUNC|O_CREAT`
flags when opening a lock file (which makes opening the lock
file substantially faster).